refactor: remove built-in vector store
