+++ compact_memory/experiment_runner.py
@@ -1,20 +1,18 @@
 from __future__ import annotations
 
-import tempfile
 import time
-from dataclasses import dataclass
 from pathlib import Path
-from typing import Optional, Dict, Any
+from typing import Dict, Any
 
 from .compression.strategies_abc import CompressionStrategy
 
 from .agent import Agent
-from .json_npy_store import JsonNpyVectorStore
+from .vector_store import InMemoryVectorStore
 from .active_memory_manager import ActiveMemoryManager
-from .chunker import Chunker, SentenceWindowChunker
-from .memory_creation import MemoryCreator, ExtractiveSummaryCreator
+from .chunker import SentenceWindowChunker
+from .memory_creation import ExtractiveSummaryCreator
 from .experiments.config import ExperimentConfig
-from .embedding_pipeline import embed_text, get_embedding_dim
+from .embedding_pipeline import get_embedding_dim
 
 
 def run_experiment(
@@ -23,16 +21,11 @@ def run_experiment(
 ) -> Dict[str, Any]:
     """Ingest ``config.dataset`` and return metrics."""
 
-    work = config.work_dir or Path(tempfile.mkdtemp())
     dim = get_embedding_dim()
-    store = JsonNpyVectorStore(
-        path=str(work), embedding_model="experiment", embedding_dim=dim
-    )
+    store = InMemoryVectorStore(embedding_dim=dim)
     if config.active_memory_params:
         store.meta.update(config.active_memory_params)
-    params = {
-        k: v for k, v in store.meta.items() if k.startswith("config_")
-    }
+    params = {k: v for k, v in store.meta.items() if k.startswith("config_")}
     if config.active_memory_params:
         params.update(config.active_memory_params)
     ActiveMemoryManager(**params)
@@ -40,8 +33,9 @@ def run_experiment(
         store,
         chunker=config.chunker or SentenceWindowChunker(),
         similarity_threshold=config.similarity_threshold,
-        summary_creator=config.summary_creator
-        or ExtractiveSummaryCreator(max_words=25),
+        summary_creator=(
+            config.summary_creator or ExtractiveSummaryCreator(max_words=25)
+        ),
     )
 
     text = Path(config.dataset).read_text()
