+++ compact_memory/prototype_system_strategy.py
@@ -12,7 +12,8 @@ from typing import Callable, Dict, List, Optional
 import numpy as np
 
 from .chunker import Chunker, SentenceWindowChunker
-from .json_npy_store import JsonNpyVectorStore, BeliefPrototype, RawMemory
+from .vector_store import VectorStore
+from .models import BeliefPrototype, RawMemory
 from .memory_creation import ExtractiveSummaryCreator, MemoryCreator
 from .prototype.canonical import render_five_w_template
 from .compression.strategies_abc import CompressedMemory, CompressionStrategy
@@ -61,7 +62,7 @@ class PrototypeSystemStrategy(CompressionStrategy):
 
     def __init__(
         self,
-        store: JsonNpyVectorStore,
+        store: VectorStore,
         *,
         chunker: Optional[Chunker] = None,
         similarity_threshold: float = 0.8,
