+++ compact_memory/response_experiment.py
@@ -11,7 +11,7 @@ import yaml
 
 from .agent import Agent
 from .active_memory_manager import ActiveMemoryManager, ConversationTurn
-from .json_npy_store import JsonNpyVectorStore
+from .vector_store import InMemoryVectorStore
 from .embedding_pipeline import embed_text, get_embedding_dim, MockEncoder
 from .chunker import SentenceWindowChunker
 from .registry import get_validation_metric_class
@@ -34,7 +34,7 @@ def _load_dataset(path: Path) -> List[Dict[str, Any]]:
 
 
 def _simple_f1(reference: str, prediction: str) -> float:
-    ref = reference.split()
+    ref = reference.split
     pred = prediction.split()
     if not ref or not pred:
         return 0.0
@@ -55,12 +55,8 @@ def _evaluate_sample(
     strategy: CompressionStrategy | None,
 ) -> Dict[str, Any]:
     mgr = ActiveMemoryManager(**params)
-
-    work_dir = tempfile.mkdtemp()
     dim = get_embedding_dim()
-    store = JsonNpyVectorStore(
-        path=work_dir, embedding_model="experiment", embedding_dim=dim
-    )
+    store = InMemoryVectorStore(embedding_dim=dim)
     agent = Agent(store, chunker=SentenceWindowChunker())
 
     for turn in sample["turns"]:
