+++ compact_memory/utils.py
@@ -4,7 +4,7 @@ import os
 from pathlib import Path
 from typing import TYPE_CHECKING, Iterable, List
 
-from .json_npy_store import JsonNpyVectorStore
+
 from .chunker import SentenceWindowChunker, _CHUNKER_REGISTRY
 from .embedding_pipeline import get_embedding_dim, EmbeddingDimensionMismatchError
 
@@ -26,29 +26,10 @@ if TYPE_CHECKING:  # pragma: no cover - for type hints only
     from .agent import Agent
 
 
-def load_agent(path: Path) -> Agent:
-    """Return an :class:`Agent` loaded from ``path``.
-
-    If the stored embedding dimension does not match the current model,
-    the store is re-initialized with the correct dimension.
-    """
-    from .agent import Agent
-
-    try:
-        store = JsonNpyVectorStore(path=str(path))
-    except FileNotFoundError as exc:
-        raise FileNotFoundError(
-            f"Agent directory '{path}' not found or is invalid"
-        ) from exc
-    except EmbeddingDimensionMismatchError:
-        from .embedding_pipeline import get_embedding_dim
-        dim = get_embedding_dim()
-        store = JsonNpyVectorStore(path=str(path), embedding_dim=dim)
-
-    chunker_id = store.meta.get("chunker", "sentence_window")
-    chunker_cls = _CHUNKER_REGISTRY.get(chunker_id, SentenceWindowChunker)
-    tau = float(store.meta.get("tau", 0.8))
-    return Agent(store, chunker=chunker_cls(), similarity_threshold=tau)
+def load_agent(path: Path) -> "Agent":
+    raise RuntimeError(
+        "Persistent storage support was removed. Provide your own loader."
+    )
 
 
 def format_ingest_results(
